# config/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Customer Management API",
    "version": "1.0.0",
    "description": "This API provides endpoints to manage customer data, including creating, updating, retrieving, and deleting customer information and associated orders.",
    "contact": {
      "name": "API Support",
      "email": "fiorezarn@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/customers/{id}/addresses": {
      "get": {
        "tags": ["Customer"],
        "summary": "Get Customer Address",
        "description": "Get the address of a specific customer by their ID",
        "operationId": "getCustomerAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the customer whose address is to be retrieved",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success - Customer address found",
            "content": {
              "application/json": {
                "example": {
                  "message": "success get customer address",
                  "data": {
                    "id": "12345",
                    "addresses": [
                      {
                        "street": "cattleya residence",
                        "city": "Tangerang Selatan",
                        "postalCode": "15418"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Customer or address not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "customer or address not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Unexpected issue",
            "content": {
              "application/json": {
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Customer"],
        "summary": "Create Address",
        "description": "Create an address for a specific customer",
        "operationId": "createAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the customer to create the address for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "city": {
                    "type": "string",
                    "example": "Tangerang Selatan"
                  },
                  "postalCode": {
                    "type": "string",
                    "example": "15418"
                  },
                  "street": {
                    "type": "string",
                    "example": "cattleya residence"
                  }
                }
              },
              "examples": {
                "Create Address": {
                  "value": {
                    "city": "Tangerang Selatan",
                    "postalCode": "15418",
                    "street": "cattleya residence"
                  }
                },
                "Missing City": {
                  "value": {
                    "postalCode": "15418",
                    "street": "cattleya residence"
                  }
                },
                "Invalid Postal Code": {
                  "value": {
                    "city": "Tangerang Selatan",
                    "postalCode": "abc123",
                    "street": "cattleya residence"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address successfully created",
            "content": {
              "application/json": {
                "example": {
                  "message": "success create address",
                  "data": {
                    "id": "12345",
                    "city": "Tangerang Selatan",
                    "postalCode": "15418",
                    "street": "cattleya residence"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing or invalid data",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid input data"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Customer not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "customer not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Unexpected issue",
            "content": {
              "application/json": {
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
    "/address/{id}": {
      "delete": {
        "tags": ["Address"],
        "summary": "Delete Address",
        "description": "Delete an address by ID",
        "operationId": "deleteAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the address to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Address deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Address deleted successfully."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Address not found. The specified ID does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Address not found"
                    },
                    "message": {
                      "type": "string",
                      "example": "The address with the specified ID does not exist."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. There was an error deleting the address data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "There was an error deleting the address data."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Address"],
        "summary": "Update Address",
        "description": "Update an address by ID",
        "operationId": "updateAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the address to update"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "city": {
                    "type": "string",
                    "example": "Tangerang Selatan"
                  },
                  "custId": {
                    "type": "number",
                    "example": 2
                  },
                  "postalCode": {
                    "type": "string",
                    "example": "15418"
                  },
                  "street": {
                    "type": "string",
                    "example": "cattleya mawar"
                  }
                }
              },
              "examples": {
                "Update Address": {
                  "value": {
                    "city": "Tangerang Selatan",
                    "custId": 2,
                    "postalCode": "15418",
                    "street": "cattleya mawar"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Address updated successfully."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Address not found. The specified ID does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Address not found"
                    },
                    "message": {
                      "type": "string",
                      "example": "The address with the specified ID does not exist."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. There was an error updating the address data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "There was an error updating the address data."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}